Minor optimizations
